requirements.txt
pyyaml
Faker
mysql
mysql-connector
mysqlclient
psycopg2-binary
python-dateutil
six
text-unidecode
python-slugify